fix: Json message parsing

+++ src/main/java/capchelin/loggingManagementSystem/documents/RxInfo.java
@@ -1,11 +1,14 @@
 package capchelin.loggingManagementSystem.documents;
 
+import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
+
+//@JsonIgnoreProperties(ignoreUnknown = true)
 public class RxInfo {
     private String gatewayID;
     private String uplinkID;
     private String name;
     private int rssi;
-    private int loRaSNR;
+    private double loRaSNR;
     private Location location;
 }
